+++ TIFoundationUtils/Cancellables/Sources/BaseCancellableBag.swift
@@ -20,8 +20,33 @@
 //  THE SOFTWARE.
 //
 
-@available(iOS 13.0.0, *)
-public protocol AsyncErrorHandler: AsyncEventHandler where EventType: Error {}
+open class BaseCancellableBag: BaseCancellable {
+    public var cancellables: [Cancellable] = []
 
-@available(iOS 13.0.0, *)
-extension AnyAsyncEventHandler: AsyncErrorHandler where EventType: Error {}
+    public override init() {}
+
+    public override func cancel() {
+        guard !isCancelled else {
+            return
+        }
+
+        cancellables.forEach { $0.cancel() }
+        cancellables.removeAll()
+
+        super.cancel()
+    }
+
+    public func add(cancellable: Cancellable?) {
+        guard let cancellable = cancellable else {
+            return
+        }
+
+        cancellables.append(cancellable)
+    }
+}
+
+public extension Cancellable {
+    func add(to cancellableBag: BaseCancellableBag) {
+        cancellableBag.add(cancellable: self)
+    }
+}