+++ TIFoundationUtils/Cancellables/Sources/NonCancellable.swift
@@ -1,5 +1,5 @@
 //
-//  Copyright (c) 2020 Touch Instinct
+//  Copyright (c) 2022 Touch Instinct
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the Software), to deal
@@ -20,23 +20,12 @@
 //  THE SOFTWARE.
 //
 
-import Foundation
+public struct NonCancellable: Cancellable {
+    public let isCancelled = true
 
-@available(iOS 11.0, *)
-open class ArchiverKeyValueEncoder: CodableKeyValueEncoder {
     public init() {}
 
-    open func encodeEncodable<Value: Encodable>(value: Value, for key: StorageKey<Value>) throws -> Data {
-        let archiver = NSKeyedArchiver(requiringSecureCoding: true)
-
-        do {
-            try archiver.encodeEncodable(value, forKey: key.rawValue)
-        } catch {
-            throw StorageError.unableToEncode(underlyingError: error)
-        }
-
-        archiver.finishEncoding()
-
-        return archiver.encodedData
+    public func cancel() {
+        // nothing
     }
 }